feat: Merges dbfsqueries.py into dbfs.py
  - Places all the code looking for DBFS references into the same source code file.
  - Places test code for DBFS references into the same source file
  - Adds more tests for non DBFS queries to make sure they do not trigger Deprecations
  - Corrects use of backticks in one test that was missed
  - Changes SQL dialect from spark to databricks
