+++ src/databricks/labs/ucx/source_code/dbfs.py
@@ -1,6 +1,9 @@
 import ast
 from collections.abc import Iterable
 
+import sqlglot
+from sqlglot.expressions import Table
+
 from databricks.labs.ucx.source_code.base import Advice, Linter, Advisory, Deprecation
 
 
@@ -72,3 +75,36 @@ class DBFSUsageLinter(Linter):
         visitor = DetectDbfsVisitor()
         visitor.visit(tree)
         return visitor.get_advices()
+
+
+class FromDbfsFolder(Linter):
+    def __init__(self):
+        self._dbfs_prefixes = ["/dbfs/mnt", "dbfs:/", "/mnt/", "/dbfs/", "/"]
+
+    @staticmethod
+    def name() -> str:
+        return 'dbfs-query'
+
+    def lint(self, code: str) -> Iterable[Advice]:
+        for statement in sqlglot.parse(code, dialect='databricks'):
+            if not statement:
+                continue
+            for table in statement.find_all(Table):
+                # Check table names for deprecated DBFS table names
+                yield from self._check_dbfs_folder(table)
+
+    def _check_dbfs_folder(self, table: Table) -> Iterable[Advice]:
+        """
+        Check if the table is a DBFS table or reference in some way
+        and yield a deprecation message if it is
+        """
+        if any(table.name.startswith(prefix) for prefix in self._dbfs_prefixes):
+            yield Deprecation(
+                code='dbfs-query',
+                message=f"The use of DBFS is deprecated: {table.name}",
+                # SQLGlot does not propagate tokens yet. See https://github.com/tobymao/sqlglot/issues/3159
+                start_line=0,
+                start_col=0,
+                end_line=0,
+                end_col=1024,
+            )
