+++ src/databricks/labs/ucx/source_code/languages.py
@@ -4,18 +4,17 @@ from databricks.labs.ucx.hive_metastore.migration_status import MigrationIndex
 from databricks.labs.ucx.source_code.base import Fixer, Linter, SequentialLinter
 from databricks.labs.ucx.source_code.pyspark import SparkSql
 from databricks.labs.ucx.source_code.queries import FromTable
-from databricks.labs.ucx.source_code.dbfsqueries import FromDbfsFolder
-from databricks.labs.ucx.source_code.dbfs import DBFSUsageLinter
+from databricks.labs.ucx.source_code.dbfs import DBFSUsageLinter, FromDbfsFolder
 
 
 class Languages:
     def __init__(self, index: MigrationIndex):
         self._index = index
         from_table = FromTable(index)
-        dbfs_from_table = FromDbfsFolder()
+        dbfs_from_folder = FromDbfsFolder()
         self._linters = {
             Language.PYTHON: SequentialLinter([SparkSql(from_table, index), DBFSUsageLinter()]),
-            Language.SQL: SequentialLinter([from_table, dbfs_from_table]),
+            Language.SQL: SequentialLinter([from_table, dbfs_from_folder]),
         }
         self._fixers: dict[Language, list[Fixer]] = {
             Language.PYTHON: [SparkSql(from_table, index)],
